Multimodal_Autism_Screening_System/Backend/train_simple.py
import os
import pickle
import numpy as np
from sklearn.preprocessing import StandardScaler
from xgboost import XGBClassifier
import tensorflow as tf
from tensorflow import keras

print("Creating directories...")
os.makedirs('saved_models', exist_ok=True)

print("Creating dummy data...")
# Create dummy training data
X = np.random.randn(100, 20)
y = np.random.randint(0, 2, 100)

print("Training Behavioral Model...")
scaler = StandardScaler()
X_scaled = scaler.fit_transform(X)
behavioral_model = XGBClassifier(n_estimators=50)
behavioral_model.fit(X_scaled, y)

# Save model
with open('saved_models/behavioral_model.pkl', 'wb') as f:
    pickle.dump({'model': behavioral_model, 'scaler': scaler}, f)
print("✓ Behavioral model saved")

print("Training Voice Model...")
X_voice = np.random.randn(100, 500, 40, 1)
y_cat = keras.utils.to_categorical(y, 2)

voice_model = keras.Sequential([
    keras.layers.Conv2D(32, (3, 3), activation='relu', input_shape=(500, 40, 1)),
    keras.layers.MaxPooling2D((2, 2)),
    keras.layers.Conv2D(64, (3, 3), activation='relu'),
    keras.layers.MaxPooling2D((2, 2)),
    keras.layers.Flatten(),
    keras.layers.Dense(32, activation='relu', name='embedding'),
    keras.layers.Dense(2, activation='softmax')
])
voice_model.compile(optimizer='adam', loss='categorical_crossentropy', metrics=['accuracy'])
voice_model.fit(X_voice, y_cat, epochs=5, verbose=1)
voice_model.save('saved_models/voice_model.h5')
print("✓ Voice model saved")

print("Training Facial Model...")
X_facial = np.random.randn(100, 224, 224, 3)

facial_model = keras.Sequential([
    keras.layers.Conv2D(32, (3, 3), activation='relu', input_shape=(224, 224, 3)),
    keras.layers.MaxPooling2D((2, 2)),
    keras.layers.Conv2D(64, (3, 3), activation='relu'),
    keras.layers.MaxPooling2D((2, 2)),
    keras.layers.Flatten(),
    keras.layers.Dense(64, activation='relu', name='embedding'),
    keras.layers.Dense(2, activation='softmax')
])
facial_model.compile(optimizer='adam', loss='categorical_crossentropy', metrics=['accuracy'])
facial_model.fit(X_facial, y_cat, epochs=5, verbose=1)
facial_model.save('saved_models/facial_model.h5')
print("✓ Facial model saved")

print("Training Fusion Model...")
# Create dummy fusion data
X_fusion = np.random.randn(100, 98)  # 2 + 32 + 64 embeddings
fusion_model = keras.Sequential([
    keras.layers.Dense(64, activation='relu', input_shape=(98,)),
    keras.layers.Dense(32, activation='relu'),
    keras.layers.Dense(2, activation='softmax')
])
fusion_model.compile(optimizer='adam', loss='categorical_crossentropy', metrics=['accuracy'])
fusion_model.fit(X_fusion, y_cat, epochs=5, verbose=1)
fusion_model.save('saved_models/fusion_model.h5')
print("✓ Fusion model saved")

print("\n✅ All models trained and saved!")
