Multimodal_Autism_Screening_System/Backend/requirements.txt
# Flask Backend Requirements
Flask==2.3.0
Flask-CORS==4.0.0
werkzeug==2.3.0

# Deep Learning
tensorflow==2.13.0
torch==2.0.0
torchvision==0.15.0

# Machine Learning
scikit-learn==1.3.0
xgboost==1.7.6

# Audio Processing
librosa==0.10.0
soundfile==0.12.1

# Image Processing
opencv-python==4.8.0
Pillow==10.0.0

# Data Processing
pandas==2.0.3
numpy==1.24.3
matplotlib==3.7.2
seaborn==0.12.2

# API and Utils
python-dotenv==1.0.0
gunicorn==21.2.0
